Initial migration to private DDN
I've decided in this commit to move all of the telco data and metadata into an isolated directory.
After working with this for a while, I think the value of shared directories for shared metadata is low,
so it makes more sense to keep each in an individual location.

Also within this, I've taken out all region config for connectors since we'll push to a single region for demos.
My view on this is that we'll have one project for each demo rather than one in each region.